--- scripts/leetcode.js
@@ -1,1291 +1,0 @@
-/* Enum for languages supported by LeetCode. */
-const languages = {
-  C: '.c',
-  'C++': '.cpp',
-  'C#': '.cs',
-  Dart: '.dart',
-  Elixir: '.ex',
-  Erlang: '.erl',
-  Go: '.go',
-  Java: '.java',
-  JavaScript: '.js',
-  Javascript: '.js',
-  Kotlin: '.kt',
-  MySQL: '.sql',
-  'MS SQL Server': '.sql',
-  Oracle: '.sql',
-  Pandas: '.py',
-  PHP: '.php',
-  Python: '.py',
-  Python3: '.py',
-  Racket: '.rkt',
-  Ruby: '.rb',
-  Rust: '.rs',
-  Scala: '.scala',
-  Swift: '.swift',
-  TypeScript: '.ts',
-};
-
-/* Commit messages */
-const readmeMsg = 'Create README - LeetHub';
-const updateReadmeMsg = 'Update README - Topic Tags';
-const discussionMsg = 'Prepend discussion post - LeetHub';
-const createNotesMsg = 'Attach NOTES - LeetHub';
-
-// problem types
-const NORMAL_PROBLEM = 0;
-const EXPLORE_SECTION_PROBLEM = 1;
-
-const WAIT_FOR_GITHUB_API_TO_NOT_THROW_409_MS = 100;
-
-/* Difficulty of most recenty submitted question */
-let difficulty = '';
-
-/* state of upload for progress */
-let uploadState = { uploading: false };
-
-const getPath = (problem, filename) => {
-  return filename ? `${problem}/${filename}` : problem;
-};
-
-/* Main function for uploading code to GitHub repo, and callback cb is called if success */
-const upload = (token, hook, content, problem, filename, sha, message) => {
-  const path = getPath(problem, filename);
-  const URL = `https://api.github.com/repos/${hook}/contents/${path}`;
-
-  let data = {
-    message,
-    content,
-    sha,
-  };
-
-  data = JSON.stringify(data);
-
-  let options = {
-    method: 'PUT',
-    headers: {
-      Authorization: `token ${token}`,
-      Accept: 'application/vnd.github.v3+json',
-    },
-    body: data,
-  };
-  let newSha;
-
-  return fetch(URL, options)
-    .then(res => {
-      if (res.status === 200 || res.status === 201) {
-        return res.json();
-      } else if (res.status === 409) {
-        throw new LeetHubError('File already exists with GitHub. Updating instead.');
-      }
-    })
-    .then(async body => {
-      newSha = body.content.sha;
-      stats = await getAndInitializeStats(problem);
-      stats.shas[problem][filename] = newSha;
-      return chrome.storage.local.set({ stats });
-    })
-    .then(() => console.log(`Successfully committed ${getPath(problem, filename)} to github`));
-};
-
-// Returns stats object. If it didn't exist, initializes stats with default difficulty values and initializes the sha object for problem
-const getAndInitializeStats = problem => {
-  return chrome.storage.local.get('stats').then(({ stats }) => {
-    if (stats == null || isEmpty(stats)) {
-      stats = {};
-      stats.shas = {};
-      stats.solved = 0;
-      stats.easy = 0;
-      stats.medium = 0;
-      stats.hard = 0;
-    }
-
-    if (stats.shas[problem] == null) {
-      stats.shas[problem] = {};
-    }
-
-    return stats;
-  });
-};
-
-const incrementStats = () => {
-  return chrome.storage.local
-    .get('stats')
-    .then(({ stats }) => {
-      stats.solved += 1;
-      stats.easy += difficulty === 'Easy' ? 1 : 0;
-      stats.medium += difficulty === 'Medium' ? 1 : 0;
-      stats.hard += difficulty === 'Hard' ? 1 : 0;
-      chrome.storage.local.set({ stats });
-      return stats
-    })
-    .then(uploadPersistentStats)
-    .catch(console.error)
-};
-
-const checkAlreadyCompleted = problemName => {
-  return chrome.storage.local.get('stats').then(({ stats }) => {
-    if (stats?.shas?.[problemName] == null) {
-      return false;
-    }
-    return true;
-  });
-};
-
-/* Main function for updating code on GitHub Repo */
-/* Read from existing file on GitHub */
-/* Discussion posts prepended at top of README */
-/* Future implementations may require appending to bottom of file */
-const update = (
-  token,
-  hook,
-  addition,
-  directory,
-  filename,
-  commitMsg,
-  shouldPreprendDiscussionPosts
-) => {
-  let responseSHA;
-
-  return getGitHubFile(token, hook, directory, filename)
-    .then(resp => resp.json())
-    .then(data => {
-      responseSHA = data.sha;
-      return decodeURIComponent(escape(atob(data.content)));
-    })
-    .then(existingContent =>
-      // https://web.archive.org/web/20190623091645/https://monsur.hossa.in/2012/07/20/utf-8-in-javascript.html
-      // In order to preserve mutation of the data, we have to encode it, which is usually done in base64.
-      // But btoa only accepts ASCII 7 bit chars (0-127) while Javascript uses 16-bit minimum chars (0-65535).
-      // EncodeURIComponent converts the Unicode Points UTF-8 bits to hex UTF-8.
-      // Unescape converts percent-encoded hex values into regular ASCII (optional; it shrinks string size).
-      // btoa converts ASCII to base64.
-      shouldPreprendDiscussionPosts
-        ? btoa(unescape(encodeURIComponent(addition + existingContent)))
-        : btoa(unescape(encodeURIComponent(existingContent)))
-    )
-    .then(newContent =>
-      upload(token, hook, newContent, directory, filename, responseSHA, commitMsg)
-    );
-};
-
-function uploadGit(
-  code,
-  problemName,
-  fileName,
-  commitMsg,
-  action,
-  shouldPrependDiscussionPosts = false
-) {
-  let token;
-  let hook;
-
-  return chrome.storage.local
-    .get(['leethub_token', 'mode_type', 'leethub_hook', 'stats'])
-    .then(({ leethub_token, leethub_hook, mode_type, stats }) => {
-      token = leethub_token;
-      if (leethub_token == undefined) {
-        throw new LeetHubError('LeethubTokenUndefined');
-      }
-
-      if (mode_type !== 'commit') {
-        throw new LeetHubError('LeetHubNotAuthorizedByGit');
-      }
-
-      hook = leethub_hook;
-      if (!hook) {
-        throw new LeetHubError('NoRepoDefined');
-      }
-
-      if (action === 'upload') {
-        /* Get SHA, if it exists */
-        const sha =
-          stats?.shas?.[problemName]?.[fileName] !== undefined
-            ? stats.shas[problemName][fileName]
-            : '';
-
-        return upload(token, hook, code, problemName, fileName, sha, commitMsg);
-      } else if (action === 'update') {
-        return update(
-          token,
-          hook,
-          code,
-          problemName,
-          fileName,
-          commitMsg,
-          shouldPrependDiscussionPosts
-        );
-      }
-    })
-    .catch(err => {
-      if (err instanceof LeetHubError) {
-        return getGitHubFile(token, hook, problemName, fileName).then(resp => resp.json());
-      } else {
-        throw err;
-      }
-    })
-    .then(data =>
-      data != null // if it isn't null, then we didn't upload successfully the first time, and must have retrieved new data and reuploaded
-        ? upload(token, hook, code, problemName, fileName, data.sha, commitMsg)
-        : undefined
-    )
-    .catch(e => console.error(new LeetHubError(e.message)));
-}
-
-/* Returns GitHub data for the file specified by `${directory}/${filename}` path */
-async function getGitHubFile(token, hook, directory, filename) {
-  const path = getPath(directory, filename);
-  const URL = `https://api.github.com/repos/${hook}/contents/${path}`;
-
-  let options = {
-    method: 'GET',
-    headers: {
-      Authorization: `token ${token}`,
-      Accept: 'application/vnd.github.v3+json',
-    },
-  };
-
-  return fetch(URL, options)
-}
-
-// Updates or creates the persistent stats from local stats
-async function uploadPersistentStats(localStats) {
-  const pStats = {leetcode: localStats}
-  const pStatsEncoded = btoa(unescape(encodeURIComponent(JSON.stringify(pStats))));
-  return delay(uploadGit, WAIT_FOR_GITHUB_API_TO_NOT_THROW_409_MS, pStatsEncoded, 'stats.json', '', `Updated stats`, 'upload')
-}
-
-// Delays `func` invocation with `...args` until after `wait` milliseconds
-function delay(func, wait, ...args) {
-  return setTimeout(() => func(...args), wait)
-}
-
-/* Checks if an elem/array exists and has length */
-function checkElem(elem) {
-  return elem && elem.length > 0;
-}
-
-function convertToSlug(string) {
-  const a = 'àáâäæãåāăąçćčđďèéêëēėęěğǵḧîïíīįìłḿñńǹňôöòóœøōõőṕŕřßśšşșťțûüùúūǘůűųẃẍÿýžźż·/_,:;';
-  const b = 'aaaaaaaaaacccddeeeeeeeegghiiiiiilmnnnnoooooooooprrsssssttuuuuuuuuuwxyyzzz------';
-  const p = new RegExp(a.split('').join('|'), 'g');
-
-  return string
-    .toString()
-    .toLowerCase()
-    .replace(/\s+/g, '-') // Replace spaces with -
-    .replace(p, c => b.charAt(a.indexOf(c))) // Replace special characters
-    .replace(/&/g, '-and-') // Replace & with 'and'
-    .replace(/[^\w\-]+/g, '') // Remove all non-word characters
-    .replace(/\-\-+/g, '-') // Replace multiple - with single -
-    .replace(/^-+/, '') // Trim - from start of text
-    .replace(/-+$/, ''); // Trim - from end of text
-}
-
-function addLeadingZeros(title) {
-  const maxTitlePrefixLength = 4;
-  var len = title.split('-')[0].length;
-  if (len < maxTitlePrefixLength) {
-    return '0'.repeat(4 - len) + title;
-  }
-  return title;
-}
-
-function formatStats(time, timePercentile, space, spacePercentile) {
-  return `Time: ${time} (${timePercentile}%), Space: ${space} (${spacePercentile}%) - LeetHub`;
-}
-
-/* Discussion Link - When a user makes a new post, the link is prepended to the README for that problem.*/
-document.addEventListener('click', event => {
-  const element = event.target;
-  const oldPath = window.location.pathname;
-
-  /* Act on Post button click */
-  /* Complex since "New" button shares many of the same properties as "Post button */
-  if (
-    element.classList.contains('icon__3Su4') ||
-    element.parentElement.classList.contains('icon__3Su4') ||
-    element.parentElement.classList.contains('btn-content-container__214G') ||
-    element.parentElement.classList.contains('header-right__2UzF')
-  ) {
-    setTimeout(function () {
-      /* Only post if post button was clicked and url changed */
-      if (
-        oldPath !== window.location.pathname &&
-        oldPath === window.location.pathname.substring(0, oldPath.length) &&
-        !Number.isNaN(window.location.pathname.charAt(oldPath.length))
-      ) {
-        const date = new Date();
-        const currentDate = `${date.getDate()}/${date.getMonth()}/${date.getFullYear()} at ${date.getHours()}:${date.getMinutes()}`;
-        const addition = `[Discussion Post (created on ${currentDate})](${window.location})  \n`;
-        const problemName = window.location.pathname.split('/')[2]; // must be true.
-
-        uploadGit(addition, problemName, 'README.md', discussionMsg, 'update', true);
-      }
-    }, 1000);
-  }
-});
-
-function LeetCodeV1() {
-  this.progressSpinnerElementId = 'leethub_progress_elem';
-  this.progressSpinnerElementClass = 'leethub_progress';
-  this.injectSpinnerStyle();
-}
-LeetCodeV1.prototype.init = async function () {};
-/* Function for finding and parsing the full code. */
-/* - At first find the submission details url. */
-/* - Then send a request for the details page. */
-/* - Parse the code from the html reponse. */
-/* - Parse the stats from the html response (explore section) */
-LeetCodeV1.prototype.findAndUploadCode = function (problemName, fileName, commitMsg, action) {
-  /* Get the submission details url from the submission page. */
-  let submissionURL;
-  const e = document.getElementsByClassName('status-column__3SUg');
-  if (checkElem(e)) {
-    // for normal problem submisson
-    const submissionRef = e[1].innerHTML.split(' ')[1];
-    submissionURL = 'https://leetcode.com' + submissionRef.split('=')[1].slice(1, -1);
-  } else {
-    // for a submission in explore section
-    const submissionRef = document.getElementById('result-state');
-    submissionURL = submissionRef.href;
-  }
-
-  if (submissionURL == undefined) {
-    return;
-  }
-  /* Request for the submission details page */
-  return fetch(submissionURL)
-    .then(res => {
-      if (res.status == 200) {
-        return res.text();
-      } else {
-        throw new Error('' + res.status);
-      }
-    })
-    .then(responseText => {
-      const doc = new DOMParser().parseFromString(responseText, 'text/html');
-      /* the response has a js object called pageData. */
-      /* Pagedata has the details data with code about that submission */
-      const scripts = doc.getElementsByTagName('script');
-      for (let i = 0; i < scripts.length; i++) {
-        const text = scripts[i].innerText;
-        if (text.includes('pageData')) {
-          /* Extract the full code */
-          const firstIndex = text.indexOf('submissionCode');
-          const lastIndex = text.indexOf('editCodeUrl');
-          let slicedText = text.slice(firstIndex, lastIndex);
-          /* slicedText has form "submissionCode: 'Details code'" */
-          /* Find the index of first and last single inverted coma. */
-          const firstInverted = slicedText.indexOf("'");
-          const lastInverted = slicedText.lastIndexOf("'");
-          /* Extract only the code */
-          const codeUnicoded = slicedText.slice(firstInverted + 1, lastInverted);
-          /* The code has some unicode. Replacing all unicode with actual characters */
-          const code = codeUnicoded.replace(/\\u[\dA-F]{4}/gi, function (match) {
-            return String.fromCharCode(parseInt(match.replace(/\\u/g, ''), 16));
-          });
-
-          /* For a submission in explore section we do not get probStat beforehand.
-            So, parse statistics from submisson page */
-          if (!commitMsg) {
-            slicedText = text.slice(text.indexOf('runtime'), text.indexOf('memory'));
-            const resultRuntime = slicedText.slice(
-              slicedText.indexOf("'") + 1,
-              slicedText.lastIndexOf("'")
-            );
-            slicedText = text.slice(text.indexOf('memory'), text.indexOf('total_correct'));
-            const resultMemory = slicedText.slice(
-              slicedText.indexOf("'") + 1,
-              slicedText.lastIndexOf("'")
-            );
-            commitMsg = `Time: ${resultRuntime}, Memory: ${resultMemory} - LeetHub`;
-          }
-
-          if (code != null) {
-            return uploadGit(
-              btoa(unescape(encodeURIComponent(code))),
-              problemName,
-              fileName,
-              commitMsg,
-              action,
-              false
-            );
-          }
-        }
-      }
-    });
-};
-// Returns the language extension
-LeetCodeV1.prototype.getLanguageExtension = function () {
-  const tag = [
-    ...document.getElementsByClassName('ant-select-selection-selected-value'),
-    ...document.getElementsByClassName('Select-value-label'),
-  ];
-  if (tag && tag.length > 0) {
-    for (let i = 0; i < tag.length; i += 1) {
-      const elem = tag[i].textContent;
-      if (elem !== undefined && languages[elem] !== undefined) {
-        return languages[elem];
-      }
-    }
-  }
-  return null;
-};
-/* function to get the notes if there is any
- the note should be opened atleast once for this to work
- this is because the dom is populated after data is fetched by opening the note */
-LeetCodeV1.prototype.getNotesIfAny = function () {
-  // there are no notes on expore
-  if (document.URL.startsWith('https://leetcode.com/explore/')) return '';
-
-  notes = '';
-  if (
-    checkElem(document.getElementsByClassName('notewrap__eHkN')) &&
-    checkElem(
-      document.getElementsByClassName('notewrap__eHkN')[0].getElementsByClassName('CodeMirror-code')
-    )
-  ) {
-    notesdiv = document
-      .getElementsByClassName('notewrap__eHkN')[0]
-      .getElementsByClassName('CodeMirror-code')[0];
-    if (notesdiv) {
-      for (i = 0; i < notesdiv.childNodes.length; i++) {
-        if (notesdiv.childNodes[i].childNodes.length == 0) continue;
-        text = notesdiv.childNodes[i].childNodes[0].innerText;
-        if (text) {
-          notes = `${notes}\n${text.trim()}`.trim();
-        }
-      }
-    }
-  }
-  return notes.trim();
-};
-// Returns a slugged num+title variation e.g. 0001-two-sum
-LeetCodeV1.prototype.getProblemNameSlug = function () {
-  const questionElem = document.getElementsByClassName('content__u3I1 question-content__JfgR');
-  const questionDescriptionElem = document.getElementsByClassName('question-description__3U1T');
-  let questionTitle = 'unknown-problem';
-  if (checkElem(questionElem)) {
-    let qtitle = document.getElementsByClassName('css-v3d350');
-    if (checkElem(qtitle)) {
-      questionTitle = qtitle[0].innerHTML;
-    }
-  } else if (checkElem(questionDescriptionElem)) {
-    let qtitle = document.getElementsByClassName('question-title');
-    if (checkElem(qtitle)) {
-      questionTitle = qtitle[0].innerText;
-    }
-  }
-  return addLeadingZeros(convertToSlug(questionTitle));
-};
-/* Gets the success state of the solution and updates html elements with new classes */
-LeetCodeV1.prototype.getSuccessStateAndUpdate = function () {
-  const successTag = document.getElementsByClassName('success__3Ai7');
-  const resultState = document.getElementById('result-state');
-
-  // check success state for a normal problem
-  if (
-    checkElem(successTag) &&
-    successTag[0].className === 'success__3Ai7' &&
-    successTag[0].innerText.trim() === 'Success'
-  ) {
-    console.log(successTag[0]);
-    successTag[0].classList.add('marked_as_success');
-    return true;
-  }
-  // check success state for a explore section problem
-  else if (
-    resultState &&
-    resultState.className === 'text-success' &&
-    resultState.innerText === 'Accepted'
-  ) {
-    resultState.classList.add('marked_as_success');
-    return true;
-  }
-
-  return false;
-};
-/* Parser function for time/space stats */
-LeetCodeV1.prototype.parseStats = function () {
-  const probStats = document.getElementsByClassName('data__HC-i');
-  if (!checkElem(probStats)) {
-    return null;
-  }
-  const time = probStats[0].textContent;
-  const timePercentile = probStats[1].textContent;
-  const space = probStats[2].textContent;
-  const spacePercentile = probStats[3].textContent;
-
-  return `Time: ${time} (${timePercentile}), Space: ${space} (${spacePercentile}) - LeetHub`;
-};
-/* Parser function for the question, question title, question difficulty, and tags */
-LeetCodeV1.prototype.parseQuestion = function () {
-  let questionUrl = window.location.href;
-  if (questionUrl.endsWith('/submissions/')) {
-    questionUrl = questionUrl.substring(0, questionUrl.lastIndexOf('/submissions/') + 1);
-  }
-  const questionElem = document.getElementsByClassName('content__u3I1 question-content__JfgR');
-  const questionDescriptionElem = document.getElementsByClassName('question-description__3U1T');
-  if (checkElem(questionElem)) {
-    const qbody = questionElem[0].innerHTML;
-
-    // Problem title.
-    let qtitle = document.getElementsByClassName('css-v3d350');
-    if (checkElem(qtitle)) {
-      qtitle = qtitle[0].innerHTML;
-    } else {
-      qtitle = 'unknown-problem';
-    }
-
-    // Problem difficulty, each problem difficulty has its own class.
-    const isHard = document.getElementsByClassName('css-t42afm');
-    const isMedium = document.getElementsByClassName('css-dcmtd5');
-    const isEasy = document.getElementsByClassName('css-14oi08n');
-
-    if (checkElem(isEasy)) {
-      difficulty = 'Easy';
-    } else if (checkElem(isMedium)) {
-      difficulty = 'Medium';
-    } else if (checkElem(isHard)) {
-      difficulty = 'Hard';
-    }
-    // Final formatting of the contents of the README for each problem
-    const markdown = `<h2><a href="${questionUrl}">${qtitle}</a></h2><h3>${difficulty}</h3><hr>${qbody}`;
-    return markdown;
-  } else if (checkElem(questionDescriptionElem)) {
-    let questionTitle = document.getElementsByClassName('question-title');
-    if (checkElem(questionTitle)) {
-      questionTitle = questionTitle[0].innerText;
-    } else {
-      questionTitle = 'unknown-problem';
-    }
-
-    const questionBody = questionDescriptionElem[0].innerHTML;
-    const markdown = `<h2>${questionTitle}</h2><hr>${questionBody}`;
-
-    return markdown;
-  }
-};
-/* Injects a spinner on left side to the "Run Code" button */
-LeetCodeV1.prototype.startSpinner = function () {
-  try {
-    elem = document.getElementById('leethub_progress_anchor_element');
-    if (!elem) {
-      elem = document.createElement('span');
-      elem.id = 'leethub_progress_anchor_element';
-      elem.style = 'margin-right: 20px;padding-top: 2px;';
-    }
-    elem.innerHTML = `<div id="${this.progressSpinnerElementId}" class="${this.progressSpinnerElementClass}"></div>`;
-    target = this.insertToAnchorElement(elem);
-    uploadState.uploading = true;
-  } catch (error) {
-    console.log(error);
-  }
-};
-/* Injects css style required for the upload progress indicator */
-LeetCodeV1.prototype.injectSpinnerStyle = function () {
-  const style = document.createElement('style');
-  style.textContent = `.${this.progressSpinnerElementClass} {pointer-events: none;width: 2.0em;height: 2.0em;border: 0.4em solid transparent;border-color: #eee;border-top-color: #3E67EC;border-radius: 50%;animation: loadingspin 1s linear infinite;} @keyframes loadingspin { 100% { transform: rotate(360deg) }}`;
-  document.head.append(style);
-};
-/* Inserts an anchor element that is specific to the page you are on (e.g. Explore) */
-LeetCodeV1.prototype.insertToAnchorElement = function (elem) {
-  if (document.URL.startsWith('https://leetcode.com/explore/')) {
-    action = document.getElementsByClassName('action');
-    if (
-      checkElem(action) &&
-      checkElem(action[0].getElementsByClassName('row')) &&
-      checkElem(action[0].getElementsByClassName('row')[0].getElementsByClassName('col-sm-6')) &&
-      action[0].getElementsByClassName('row')[0].getElementsByClassName('col-sm-6').length > 1
-    ) {
-      target = action[0].getElementsByClassName('row')[0].getElementsByClassName('col-sm-6')[1];
-      elem.className = 'pull-left';
-      if (target.childNodes.length > 0) target.childNodes[0].prepend(elem);
-    }
-  } else {
-    if (checkElem(document.getElementsByClassName('action__38Xc'))) {
-      target = document.getElementsByClassName('action__38Xc')[0];
-      elem.className = 'runcode-wrapper__8rXm';
-      if (target.childNodes.length > 0) target.childNodes[0].prepend(elem);
-    }
-  }
-};
-/* Creates a ✔️ tick mark before "Run Code" button signaling LeetHub has done its job */
-LeetCodeV1.prototype.markUploaded = function () {
-  elem = document.getElementById(this.progressSpinnerElementId);
-  if (elem) {
-    elem.className = '';
-    style =
-      'display: inline-block;transform: rotate(45deg);height:24px;width:12px;border-bottom:7px solid #78b13f;border-right:7px solid #78b13f;';
-    elem.style = style;
-  }
-};
-/* Creates a ❌ failed tick mark before "Run Code" button signaling that upload failed */
-LeetCodeV1.prototype.markUploadFailed = function () {
-  elem = document.getElementById(this.progressSpinnerElementId);
-  if (elem) {
-    elem.className = '';
-    style =
-      'display: inline-block;transform: rotate(45deg);height:24px;width:12px;border-bottom:7px solid red;border-right:7px solid red;';
-    elem.style = style;
-  }
-};
-LeetCodeV1.prototype.updateReadmeTopicTagsWithProblem = function () {
-  console.log(`Skipping...updating README with tags is not supported on old UI`);
-};
-
-function LeetCodeV2() {
-  this.submissionData;
-  this.submissionId;
-  this.progressSpinnerElementId = 'leethub_progress_elem';
-  this.progressSpinnerElementClass = 'leethub_progress';
-  this.injectSpinnerStyle();
-}
-LeetCodeV2.prototype.init = async function () {
-  const submissionId = this.submissionId;
-
-  // Query for getting the solution runtime and memory stats, the code, the coding language, the question id, question title and question difficulty
-  const submissionDetailsQuery = {
-    query: `\n    query submissionDetails($submissionId: Int!) {\n submissionDetails(submissionId: $submissionId) {\n runtime\n    runtimeDisplay\n    runtimePercentile\n    runtimeDistribution\n    memory\n    memoryDisplay\n    memoryPercentile\n    memoryDistribution\n    code\n    timestamp\n    statusCode\n lang {\n name\n      verboseName\n }\n question {\n questionId\n    questionFrontendId\n    title\n    titleSlug\n    content\n    difficulty\n topicTags {\n name\n    slug\n }\n }\n    notes\n    runtimeError\n  }\n}\n    `,
-    variables: { submissionId: submissionId },
-    operationName: 'submissionDetails',
-  };
-  const options = {
-    method: 'POST',
-    headers: {
-      cookie: document.cookie, // required to authorize the API request
-      'content-type': 'application/json',
-    },
-    body: JSON.stringify(submissionDetailsQuery),
-  };
-  const data = await fetch('https://leetcode.com/graphql/', options)
-    .then(res => res.json())
-    .then(res => res.data.submissionDetails);
-
-  this.submissionData = data;
-};
-LeetCodeV2.prototype.findAndUploadCode = function (problemName, fileName, commitMsg, action) {
-  const code = this.getCode();
-  if (!code) {
-    throw new LeetHubError('SolutionCodeNotFound');
-  }
-
-  return uploadGit(
-    btoa(unescape(encodeURIComponent(code))),
-    problemName,
-    fileName,
-    commitMsg,
-    action,
-    false
-  );
-};
-LeetCodeV2.prototype.getCode = function () {
-  if (this.submissionData != null) {
-    return this.submissionData.code;
-  }
-
-  const code = document.getElementsByTagName('code');
-  if (!checkElem(code)) {
-    return null;
-  }
-
-  return code[0].innerText;
-};
-LeetCodeV2.prototype.getLanguageExtension = function () {
-  if (this.submissionData != null) {
-    return languages[this.submissionData.lang.verboseName];
-  }
-
-  const tag = document.querySelector('button[id^="headlessui-listbox-button"]');
-  if (!tag) {
-    throw new LeetHubError('LanguageButtonNotFound');
-  }
-
-  const lang = tag.innerText;
-  if (languages[lang] === undefined) {
-    throw new LeetHubError(`UnknownLanguage::${lang}`);
-  }
-
-  return languages[lang];
-};
-LeetCodeV2.prototype.getNotesIfAny = function () {};
-LeetCodeV2.prototype.getProblemNameSlug = function () {
-  const slugTitle = this.submissionData.question.titleSlug;
-  const qNum = this.submissionData.question.questionId;
-
-  return addLeadingZeros(qNum + '-' + slugTitle);
-};
-LeetCodeV2.prototype.getSuccessStateAndUpdate = function () {
-  const successTag = document.querySelectorAll('[data-e2e-locator="submission-result"]');
-  if (checkElem(successTag)) {
-    console.log(successTag[0]);
-    successTag[0].classList.add('marked_as_success');
-    return true;
-  }
-  return false;
-};
-LeetCodeV2.prototype.parseStats = function () {
-  if (this.submissionData != null) {
-    const runtimePercentile =
-      Math.round((this.submissionData.runtimePercentile + Number.EPSILON) * 100) / 100;
-    const spacePercentile =
-      Math.round((this.submissionData.memoryPercentile + Number.EPSILON) * 100) / 100;
-    return formatStats(
-      this.submissionData.runtimeDisplay,
-      runtimePercentile,
-      this.submissionData.memoryDisplay,
-      spacePercentile
-    );
-  }
-
-  const probStats = document.getElementsByClassName('flex w-full pb-4')[0].innerText.split('\n');
-  if (!checkElem(probStats)) {
-    return null;
-  }
-
-  const time = probStats[1];
-  const timePercentile = probStats[3];
-  const space = probStats[5];
-  const spacePercentile = probStats[7];
-
-  return formatStats(time, timePercentile, space, spacePercentile);
-};
-LeetCodeV2.prototype.parseQuestion = function () {
-  let markdown;
-  if (this.submissionData != null) {
-    const questionUrl = window.location.href.split('/submissions')[0];
-    const qTitle = `${this.submissionData.question.questionId}. ${this.submissionData.question.title}`;
-    const qBody = this.parseQuestionDescription();
-
-    difficulty = this.submissionData.question.difficulty;
-
-    // Final formatting of the contents of the README for each problem
-    markdown = `<h2><a href="${questionUrl}">${qTitle}</a></h2><h3>${difficulty}</h3><hr>${qBody}`;
-  } else {
-    // TODO: get the README markdown via scraping. Right now this isn't possible.
-    markdown = null;
-  }
-
-  return markdown;
-};
-LeetCodeV2.prototype.parseQuestionTitle = function () {
-  if (this.submissionData != null) {
-    return this.submissionData.question.title;
-  }
-
-  let questionTitle = document
-    .getElementsByTagName('title')[0]
-    .innerText.split(' ')
-    .slice(0, -2)
-    .join(' ');
-
-  if (questionTitle === '') {
-    questionTitle = 'unknown-problem';
-  }
-
-  return questionTitle;
-};
-LeetCodeV2.prototype.parseQuestionDescription = function () {
-  if (this.submissionData != null) {
-    return this.submissionData.question.content;
-  }
-
-  const description = document.getElementsByName('description');
-  if (!checkElem(description)) {
-    return null;
-  }
-  return description[0].content;
-};
-LeetCodeV2.prototype.parseDifficulty = function () {
-  if (this.submissionData != null) {
-    return this.submissionData.question.difficulty;
-  }
-
-  const diffElement = document.getElementsByClassName('mt-3 flex space-x-4');
-  if (checkElem(diffElement)) {
-    return diffElement[0].children[0].innerText;
-  }
-  // Else, we're not on the description page. Nothing we can do.
-  return 'unknown';
-};
-LeetCodeV2.prototype.startSpinner = function () {
-  let elem = document.getElementById('leethub_progress_anchor_element');
-  if (!elem) {
-    elem = document.createElement('span');
-    elem.id = 'leethub_progress_anchor_element';
-    elem.style = 'margin-right: 20px;padding-top: 2px;';
-  }
-  elem.innerHTML = `<div id="${this.progressSpinnerElementId}" class="${this.progressSpinnerElementClass}"></div>`;
-  this.insertToAnchorElement(elem);
-  uploadState.uploading = true;
-};
-LeetCodeV2.prototype.injectSpinnerStyle = function () {
-  const style = document.createElement('style');
-  style.textContent = `.${this.progressSpinnerElementClass} {pointer-events: none;width: 2.0em;height: 2.0em;border: 0.4em solid transparent;border-color: #eee;border-top-color: #3E67EC;border-radius: 50%;animation: loadingspin 1s linear infinite;} @keyframes loadingspin { 100% { transform: rotate(360deg) }}`;
-  document.head.append(style);
-};
-LeetCodeV2.prototype.insertToAnchorElement = function (elem) {
-  if (document.URL.startsWith('https://leetcode.com/explore/')) {
-    // TODO: support spinner when answering problems on Explore pages
-    // const submitWait = document.getElementById('submit_wait')
-    // if (submitWait) {
-    //   const parent = submitWait.parentElement
-    //   elem.className = 'runcode-wrapper__8rXm';
-    //   parent.insertBefore(elem, submitWait)
-    // }
-    // Must initiate script from a different submit button
-    // document.querySelector('.btn.btn-primary.btn-round')
-    
-    // return;
-  }
-  // TODO: target within the Run and Submit div regardless of UI position of submit button
-  const target = document.querySelector('[data-e2e-locator="submission-result"]').parentElement;
-  if (target) {
-    elem.className = 'runcode-wrapper__8rXm';
-    target.appendChild(elem);
-  }
-};
-LeetCodeV2.prototype.markUploaded = function () {
-  let elem = document.getElementById(this.progressSpinnerElementId);
-  if (elem) {
-    elem.className = '';
-    style =
-      'display: inline-block;transform: rotate(45deg);height:24px;width:12px;border-bottom:7px solid #78b13f;border-right:7px solid #78b13f;';
-    elem.style = style;
-  }
-};
-LeetCodeV2.prototype.markUploadFailed = function () {
-  let elem = document.getElementById(this.progressSpinnerElementId);
-  if (elem) {
-    elem.className = '';
-    style =
-      'display: inline-block;transform: rotate(45deg);height:24px;width:12px;border-bottom:7px solid red;border-right:7px solid red;';
-    elem.style = style;
-  }
-};
-LeetCodeV2.prototype.updateReadmeTopicTagsWithProblem = async function (problemName) {
-  const { leethub_token, leethub_hook, stats } = await chrome.storage.local.get([
-    'leethub_token',
-    'leethub_hook',
-    'stats',
-  ]);
-  const resp = await getGitHubFile(leethub_token, leethub_hook, 'README.md');
-  const { content } = await resp.json();
-
-  let readme = decodeURIComponent(escape(atob(content)));
-  for (let topic of this.submissionData.question.topicTags) {
-    readme = appendProblemToTopic(topic.name, readme, leethub_hook, problemName);
-  }
-  readme = sortTopicTablesInMarkdown(readme);
-  readme = btoa(unescape(encodeURIComponent(readme)));
-  // For some reason, this API call needs a longer delay.
-  return delay(uploadGit, WAIT_FOR_GITHUB_API_TO_NOT_THROW_409_MS + 100, readme, 'README.md', '', updateReadmeMsg, 'upload');
-};
-
-// Appends a problem title to each Topic section in the README.md
-function appendProblemToTopic(topic, markdownFile, hook, problem) {
-  const url = `https://github.com/${hook}/tree/master/${problem}`;
-
-  // Check if a table already belongs to topic, or add it
-  let topicTableIndex = markdownFile.indexOf(`# ${topic}`);
-  if (topicTableIndex === -1) {
-    markdownFile += `\n# ${topic}\n|  |\n| ------- |\n`;
-  }
-
-  // Find the Topic table
-  topicTableIndex = markdownFile.lastIndexOf(`# ${topic}`);
-  const nextTableIndex = markdownFile.indexOf('# ', topicTableIndex + 1);
-  let topicTable =
-    nextTableIndex === -1
-      ? markdownFile.slice(topicTableIndex)
-      : markdownFile.slice(topicTableIndex, nextTableIndex);
-  topicTable = topicTable.trim();
-
-  // Check if the problem exists in table, prevent duplicate add
-  const problemIndex = topicTable.indexOf(problem);
-  if (problemIndex !== -1) {
-    return markdownFile;
-  }
-
-  // Append problem to the Topic
-  const newRow = `| [${problem}](${url}) |`;
-  topicTable = [topicTable, newRow, '\n'].join('\n');
-
-  // Replace the old Topic table with the updated one in the markdown file
-  markdownFile =
-    markdownFile.slice(0, topicTableIndex) +
-    topicTable +
-    (nextTableIndex === -1 ? '' : markdownFile.slice(nextTableIndex));
-
-  return markdownFile;
-}
-
-// Sorts each Topic table by the problem number
-function sortTopicTablesInMarkdown(markdownFile) {
-  let topics = markdownFile.split('# ');
-
-  // Remove the first element (empty) and next element (repo title + description)
-  topics.shift();
-  let description = topics.shift();
-
-  // Loop over each problem topic
-  topics = topics.map(section => {
-    let lines = section.trim().split('\n');
-
-    // Get the problem topic
-    let topic = lines.shift();
-
-    // Remove the header and header separator
-    lines = lines.slice(2);
-
-    lines.sort((a, b) => {
-      let numA = parseInt(a.match(/\/(\d+)-/)[1]);
-      let numB = parseInt(b.match(/\/(\d+)-/)[1]);
-      return numA - numB;
-    });
-
-    // Reconstruct the section
-    return ['# ' + topic].concat('|  |', '| ------- |', lines).join('\n');
-  });
-
-  // Reconstruct the file
-  markdownFile = ['# ' + description.trim(), '\n'].concat(topics).join('\n');
-  return markdownFile;
-}
-
-function loader(leetCode) {
-  let iterations = 0;
-  const intervalId = setInterval(async () => {
-    try {
-      const isSuccessfulSubmission = leetCode.getSuccessStateAndUpdate();
-      if (!isSuccessfulSubmission) {
-        iterations++;
-        if (iterations > 9) { // poll for max 10 attempts (10 seconds)
-          throw new LeetHubError('Could not find successful submission after 10 seconds.')
-        }
-        return;
-      }
-      leetCode.startSpinner();
-
-      // If successful, stop polling
-      clearInterval(intervalId);
-
-      // For v2, query LeetCode API for submission results
-      await leetCode.init();
-      
-      const probStats = leetCode.parseStats();
-      if (!probStats) {
-        throw new LeetHubError('SubmissionStatsNotFound');
-      }
-      
-      const probStatement = leetCode.parseQuestion();
-      if (!probStatement) {
-        throw new LeetHubError('ProblemStatementNotFound');
-      }
-      
-      const problemName = leetCode.getProblemNameSlug();
-      const alreadyCompleted = await checkAlreadyCompleted(problemName);
-      const language = leetCode.getLanguageExtension();
-      if (!language) {
-        throw new LeetHubError('LanguageNotFound');
-      }
-
-      /* Upload README */
-      const uploadReadMe = await chrome.storage.local.get('stats').then(({ stats }) => {
-        const shaExists = stats?.shas?.[problemName]?.['README.md'] !== undefined;
-
-        if (!shaExists) {
-          return uploadGit(
-            btoa(unescape(encodeURIComponent(probStatement))),
-            problemName,
-            'README.md',
-            readmeMsg,
-            'upload',
-            false
-          );
-        }
-      });
-
-      /* Upload Notes if any*/
-      notes = leetCode.getNotesIfAny();
-      let uploadNotes;
-      if (notes != undefined && notes.length > 0) {
-        uploadNotes = uploadGit(
-          btoa(unescape(encodeURIComponent(notes))),
-          problemName,
-          'NOTES.md',
-          createNotesMsg,
-          'upload',
-          false
-        );
-      }
-
-      /* Upload code to Git */
-      const uploadCode = leetCode.findAndUploadCode(
-        problemName,
-        problemName + language,
-        probStats,
-        'upload'
-      );
-
-      /* Group problem into its relevant topics */
-      const updateRepoReadMeWithTopicTag = leetCode.updateReadmeTopicTagsWithProblem(problemName);
-
-      await Promise.all([uploadReadMe, uploadNotes, uploadCode, updateRepoReadMeWithTopicTag]);
-
-      uploadState.uploading = false;
-      leetCode.markUploaded();
-
-      if (!alreadyCompleted) {
-        incrementStats(); // Increments local and persistent stats
-      }
-    } catch (err) {
-      uploadState.uploading = false;
-      leetCode.markUploadFailed();
-      clearInterval(intervalId);
-      console.log(err);
-    }
-  }, 1000);
-}
-
-function isEmpty(obj) {
-  for (const prop in obj) {
-    if (Object.hasOwn(obj, prop)) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-// Submit by Keyboard Shortcuts only support on LeetCode v2
-function wasSubmittedByKeyboard(event) {
-  const isEnterKey = event.key === 'Enter';
-  const isMacOS = window.navigator.userAgent.includes('Mac');
-
-  // Adapt to MacOS operating system
-  return isEnterKey && ((isMacOS && event.metaKey) || (!isMacOS && event.ctrlKey));
-}
-
-// Get SubmissionID by listening for URL changes to `/submissions/(d+)` format
-async function listenForSubmissionId() {
-  const { submissionId } = await chrome.runtime.sendMessage({
-    type: 'LEETCODE_SUBMISSION',
-  });
-  if (submissionId == null) {
-    console.log(new LeetHubError('SubmissionIdNotFound'));
-    return;
-  }
-  return submissionId;
-}
-
-async function v2SubmissionHandler(event, leetCode) {
-  if (event.type !== 'click' && !wasSubmittedByKeyboard(event)) {
-    return;
-  }
-
-  const authenticated = !isEmpty(await chrome.storage.local.get(['leethub_token'])) && !isEmpty(await chrome.storage.local.get(['leethub_hook']))
-  if (!authenticated) {
-    throw new LeetHubError('UserNotAuthenticated')
-  }
-
-  // is click or is ctrl enter
-  const submissionId = await listenForSubmissionId();
-  leetCode.submissionId = submissionId;
-  loader(leetCode);
-  return true;
-}
-
-// Use MutationObserver to determine when the submit button elements are loaded
-const submitBtnObserver = new MutationObserver(function (_mutations, observer) {
-  const v1SubmitBtn = document.querySelector('[data-cy="submit-code-btn"]');
-  const v2SubmitBtn = document.querySelector('[data-e2e-locator="console-submit-button"]');
-  const textareaList = document.getElementsByTagName('textarea');
-  const textarea =
-    textareaList.length === 4
-      ? textareaList[2]
-      : textareaList.length === 2
-        ? textareaList[0]
-        : textareaList[1];
-
-  if (v1SubmitBtn) {
-    observer.disconnect();
-
-    const leetCode = new LeetCodeV1();
-    v1SubmitBtn.addEventListener('click', () => loader(leetCode));
-    return;
-  }
-
-  if (v2SubmitBtn && textarea) {
-    observer.disconnect();
-
-    const leetCode = new LeetCodeV2();
-    if (!!!v2SubmitBtn.onclick) {
-      textarea.addEventListener('keydown', e => v2SubmissionHandler(e, leetCode));
-      v2SubmitBtn.onclick = e => v2SubmissionHandler(e, leetCode);
-    }
-  }
-});
-
-submitBtnObserver.observe(document.body, {
-  childList: true,
-  subtree: true,
-});
-
-/* Sync to local storage */
-chrome.storage.local.get('isSync', data => {
-  keys = [
-    'leethub_token',
-    'leethub_username',
-    'pipe_leethub',
-    'stats',
-    'leethub_hook',
-    'mode_type',
-  ];
-  if (!data || !data.isSync) {
-    keys.forEach(key => {
-      chrome.storage.sync.get(key, data => {
-        chrome.storage.local.set({ [key]: data[key] });
-      });
-    });
-    chrome.storage.local.set({ isSync: true }, () => {
-      console.log('LeetHub Synced to local values');
-    });
-  } else {
-    console.log('LeetHub Local storage already synced!');
-  }
-});
-
-class LeetHubError extends Error {
-  constructor(message) {
-    super(message);
-    this.name = 'LeetHubErr';
-  }
-}
-
-class LeetHubNetworkError extends LeetHubError {
-  constructor(response) {
-    super(response.statusText)
-    this.status = response.status
-  }
-}
-
-(showManualSubmitBtn = () => {
-  const getSubmissionPageBtns = () => {
-    return document.querySelector(
-      '.flex.flex-none.gap-2:not(.justify-center):not(.justify-between)'
-    );
-  };
-
-  const createToolTip = () => {
-    const toolTip = document.createElement('div');
-    toolTip.id = 'leethub-upload-tooltip';
-    toolTip.textContent =
-      'Manually upload this submission to GitHub (beta).\nThis will OVERWRITE your current submission.\nPlease be mindful of your GitHub rate-limits.';
-    toolTip.className =
-      'fixed bg-sd-popover text-sd-popover-foreground rounded-sd-md z-modal text-xs text-left font-normal whitespace-pre-line shadow p-3 border-sd-border border cursor-default translate-y-20 transition-opacity opacity-0 transition-delay-1000 duration-300 group-hover:opacity-100';
-    return toolTip;
-  };
-
-  const createGitIcon = () => {
-    const uploadIcon = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
-    uploadIcon.setAttribute('id', 'leethub-upload-icon');
-    uploadIcon.setAttribute('xmlns', 'http://www.w3.org/2000/svg');
-    uploadIcon.setAttribute('width', '16');
-    uploadIcon.setAttribute('height', '17');
-    uploadIcon.setAttribute('viewBox', '0 0 38.999866 56.642887');
-
-    const path = document.createElementNS('http://www.w3.org/2000/svg', 'path');
-    path.setAttribute(
-      'style',
-      'fill:#fcfcfc;fill-opacity:1;stroke:#ffffff;stroke-width:3;stroke-dasharray:none;stroke-opacity:1'
-    );
-    path.setAttribute(
-      'd',
-      'm 19.775372,2.121319 -9.072314,9.072314 a 0.51539412,0.66999737 45 0 0 -0.109554,0.838192 0.49679682,0.64582142 45 0 0 0.810286,-0.125057 l 7.846033,-7.846033 v 30.608468 a 0.47397466,0.47397466 0 0 0 0.473873,0.473873 h 0.0093 a 0.51713218,0.51713218 0 0 0 0.516765,-0.517281 V 4.018877 l 7.559745,7.560262 a 0.62190211,0.49679682 45 0 0 0.793233,0.107487 0.64518265,0.51539412 45 0 0 -0.09198,-0.820621 l -8.033101,-8.033102 0.0047,-0.0047 z m 7.81141,17.001029 v 0.999939 l 5.229655,0.01189 a 3.6922154,3.6922154 0 0 1 3.683496,3.692281 v 26.633 a 3.6835681,3.6835681 0 0 1 -3.683496,3.683496 H 6.1834371 a 3.6835681,3.6835681 0 0 1 -3.683496,-3.683496 v -26.633 a 3.6835681,3.6835681 0 0 1 3.683496,-3.683496 H 11.538666 V 19.143023 H 6.3121111 a 4.8119141,4.8119141 0 0 0 -4.812109,4.812109 v 26.375651 a 4.8119141,4.8119141 0 0 0 4.812109,4.81211 H 32.687762 a 4.8119141,4.8119141 0 0 0 4.81211,-4.81211 V 23.955128 a 4.8220648,4.8220648 0 0 0 -4.81211,-4.822444 z'
-    );
-
-    uploadIcon.appendChild(path);
-    return uploadIcon;
-  };
-
-  const addManualSubmitBtn = () => {
-    const btns = getSubmissionPageBtns();
-    if (btns.innerText.includes('Solution') && !btns.innerText.includes('LeetHub')) {
-      btns.appendChild(
-        (() => {
-          const btn = document.createElement('button');
-          btn.innerText = 'Sync w/ LeetHub';
-          btn.setAttribute('style', 'background-color:darkorange');
-          btn.setAttribute(
-            'class',
-            'group whitespace-nowrap focus:outline-none text-label-r bg-green-s dark:bg-dark-blue-s hover:bg-green-3 dark:hover:bg-dark-blue-3 flex items-center justify-center gap-2 rounded-lg px-3.5 py-1.5 text-sm font-medium'
-          );
-
-          btn.prepend(createGitIcon());
-          btn.appendChild(createToolTip());
-          btn.addEventListener(
-            'click',
-            debounce(
-              async () => {
-                // No waiting for url change. Manual submissions have submission url.
-                const authenticated = !isEmpty(await chrome.storage.local.get(['leethub_token'])) && !isEmpty(await chrome.storage.local.get(['leethub_hook']))
-                if (!authenticated) {
-                  throw new LeetHubError('UserNotAuthenticated')
-                }
-                
-                const leetCode = new LeetCodeV2();
-                const submissionId = window.location.href.match(
-                  /leetcode\.com\/.*\/submissions\/(\d+)/
-                )[1];
-                leetCode.submissionId = submissionId;
-                loader(leetCode);
-                return;
-              },
-              5000,
-              true
-            )
-          );
-          return btn;
-        })()
-      );
-    }
-  };
-
-  const submissionPageBtnsObserver = new MutationObserver((_, observer) => {
-    const url = window.location.href;
-    const btns = getSubmissionPageBtns();
-
-    if (btns && btns.children.length < 3 && url.match(/\/submissions\//)) {
-      observer.disconnect();
-      addManualSubmitBtn();
-    }
-  });
-
-  window.navigation.addEventListener('navigate', () => {
-    const isSubmissionUrl = window.location.href.match(/leetcode\.com\/(.*)\/submissions\/(\d+)/);
-    if (isSubmissionUrl) {
-      submissionPageBtnsObserver.observe(document.body, {
-        childList: true,
-        subtree: true,
-      });
-    }
-  });
-
-  // Returns a function that can be immediately invoked but will start a timeout of 'wait' milliseconds before it can be called again.
-  function debounce(func, wait, immediate) {
-    let timeout;
-    return function () {
-      const context = this, args = arguments;
-      const later = function () {
-        timeout = null;
-        if (!immediate) func.apply(context, args);
-      };
-      const callNow = immediate && !timeout;
-      clearTimeout(timeout);
-      timeout = setTimeout(later, wait);
-      if (callNow) func.apply(context, args);
-    };
-  }
-})();
